requirements.txt
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
beautifulsoup4==4.12.3
certifi==2024.2.2
exceptiongroup==1.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
lxml==5.1.0
playwright==1.41.2
pyee==11.0.1
python-dotenv==1.0.1
schedule==1.2.1
sniffio==1.3.0
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.9.0
